tworzenie generatora samochodów i wrzucenie samochdów do listy

+++ src/Mechanic.java
@@ -1,10 +1,10 @@
 public class Mechanic {
     public String name;
     public Double price;
-    public Double lucky;
-    public Double damagedProbability;
+    public Integer lucky;
+    public Integer damagedProbability;
 
-    public Mechanic(String name, Double price, Double lucky, Double damagedProbability){
+    public Mechanic(String name, Double price, Integer lucky, Integer damagedProbability){
         this.name=name;
         this.price=price;
         this.lucky=lucky;
